work with new strongly typed numpyType

+++ dotty-tensorflow/src/main/scala/me.shadaj.scalapy.tensorflow/example/BidirectionalLSTMExample.scala
@@ -2,6 +2,7 @@ package me.shadaj.scalapy.tensorflow.example
 
 import me.shadaj.scalapy.py
 import me.shadaj.scalapy.tensorflow.scala.utils.Modules._
+import me.shadaj.scalapy.tensorflow.nd2Tensor
 import Int.int2long
 import scala.language.implicitConversions
 
@@ -30,12 +31,11 @@ object BidirectionalLSTMExample extends Runnable {
     println(s"${x_test.length} test sequences")
 
     println("Pad sequences (samples x time)")
-    val x_train1 = sequence.pad_sequences(x_train, maxlen = maxlen)
-    val x_test1 = sequence.pad_sequences(x_test, maxlen = maxlen)
+    val x_train1 = sequence.pad_sequences(x_train, maxlen = maxlen).astype(np.float32)
+    val x_test1 = sequence.pad_sequences(x_test, maxlen = maxlen).astype(np.float32)
     println(s"x_train shape: ${x_train1.shape}")
     println(s"x_test shape: ${x_test1.shape}")
-    //    y_train = np.array(y_train)
-    //    y_test = np.array(y_test)
+
     val model = keras1.models.Sequential()
     model.add(layers.Embedding(max_features, 128, input_length = maxlen))
     model.add(layers.Bidirectional(layers.LSTM(64)))
